Extracted Tinybird concerns into its own file
